requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.2.0
beautifulsoup4==4.12.2
colorama==0.4.6
coverage==7.4.0
Django==5.0
django-bootstrap4==23.4
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
py==1.11.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.7.0
pytz==2023.3.post1
soupsieve==2.5
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.3
whitenoise==6.6.0
